Harmonize the structure with the TraceGateway config builder
