+++ internal/otelcollector/config/loggateway/testdata/single-pipeline.yaml
@@ -46,17 +46,9 @@ processors:
         send_batch_size: 512
         timeout: 10s
         send_batch_max_size: 512
-    memory_limiter:
-        check_interval: 1s
-        limit_percentage: 75
-        spike_limit_percentage: 15
-    transform/set-observed-time-if-zero:
-        error_mode: ignore
-        log_statements:
-            - statements:
-                - set(log.observed_time, Now())
-              conditions:
-                - log.observed_time_unix_nano == 0
+    istio_enrichment:
+        scope_version: 1.0.0
+    istio_noise_filter: {}
     k8sattributes:
         auth_type: serviceAccount
         passthrough: false
@@ -98,7 +90,14 @@ processors:
                   name: k8s.pod.uid
             - sources:
                 - from: connection
-    istio_noise_filter: {}
+    memory_limiter:
+        check_interval: 1s
+        limit_percentage: 75
+        spike_limit_percentage: 15
+    resource/drop-kyma-attributes:
+        attributes:
+            - action: delete
+              pattern: kyma.*
     resource/insert-cluster-attributes:
         attributes:
             - action: insert
@@ -113,12 +112,13 @@ processors:
         resource_attributes:
             - kyma.kubernetes_io_app_name
             - kyma.app_name
-    resource/drop-kyma-attributes:
-        attributes:
-            - action: delete
-              pattern: kyma.*
-    istio_enrichment:
-        scope_version: 1.0.0
+    transform/set-observed-time-if-zero:
+        error_mode: ignore
+        log_statements:
+            - statements:
+                - set(log.observed_time, Now())
+              conditions:
+                - log.observed_time_unix_nano == 0
 exporters:
     otlp/test:
         endpoint: ${OTLP_ENDPOINT_TEST}
